op40-move: extract allocation_sizing.py from common.py
